+++ android/src/main/kotlin/com/flutter/background/activity/launch/flutter_alarm_background_trigger/db/AlarmDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.flutter.background.activity.launch.flutter_alarm_background_trigger.db.coverters.StatusTypeConverter
 
-@Database(entities = [AlarmItem::class], version = 3)
+@Database(entities = [AlarmItem::class], version = 3, exportSchema = true)
 @TypeConverters(StatusTypeConverter::class)
 abstract class AlarmDatabase : RoomDatabase() {
     abstract fun alarmsDao(): AlarmDao?